feat: large rewrite of the search functionality
the searching implementation was largely rewritten!

with this change, searchers are now structured units with a
common interface that wrap the underlying tantivy implementation

when a search unit is created for a certain channel, a
timer can be started that starts a thread in the background which
periodically fetches the newest commit from github.
if it detects a new commit, the channel's options (and now also packages)
are rebuilt and on success replaced in the underlying data structure,
cached to disk and indexed for search.
the server backend implmentation holds a reference to the shared options
data structure, so that the newest results can be provided for each
request

when updating a searcher's entries, the old ones are also deleted now,
such that removed options cannot be searched in order to correctly
reflect the correct branch state. previously this was not the case.

this commit also introduces a new package search for fcio packages
as well as upstream nixos packages for a given fc channel.
the packages are generated together with the options and the
webinterface exposes the package search as `/search/packages`

in order to accurately reflect the scale of these changes, the crate
version was incremented from 0.1 to 0.2.