fix: reolad newest values after commit (update)
this commit fixes a bug where new updates to search values
(packages + options) were not reflected at runtime and
required a restart.
this was due to the fact that at channelsearcher initialisation the
reader was dropped and an old version of the searcher store instead

since the searcher needs to be reinitialized for every query to
guarantee up-to-date results, this lead to a non-functional search when
the searcher is indexed first.

by storing the reader instead and reinitializing the searcher on every
request, new values are immediately reflected in search results