feat: [options] tweak scoring + set prefix search
when searching, especially with active search, input should be
considered a prefix of the intended search.
to reflect that in the search results, the name field was changed to
consider searches on it prefixes of indexed values.

additionally the scoring of the search results was adjusted to highlight
flying circus options and to add a small boost to `enable` options.