feat: improve shared data structure peformance
The data that is synced across threads for incoming requests needs to
be stored in a data structure that facilitates udpating the underlying
data at runtime so that up-to-date search results can be served.

The previous implementation utilized Mutexes to acquire read access for
searching and write access to update the underlying data structure.
With an optimized data structure for frequent reads and infrequent
writes the performance can be improved significantly while eliminating
cross-thread syncronization slowdowns.

This commit also introduces some profiling tools for the development
environment provided in the nix flake.