fix: actually update the channels
Previously, only a copy of the channel searchers was actually updated as
well as the on disk data and the tantivy searchers.
Since the underlying tantivy searcher was updated but not the lookup-table
in the searchers which is used to correlate the returned documents from
the tantivy searcher back to options and packages, changes to options
and packages would return documents from tantivy that the searcher
structure could not correlate back to actual options.